Add confirmation dialog when cancelling exercise creation
